fix model, replace Integer to BigInteger

+++ app/db/models/users.py
@@ -18,7 +18,7 @@ class User(Base):
 
     id = sa.Column(sa.Integer, primary_key=True, index=True, autoincrement=True)
 
-    telegram_id = sa.Column(sa.Integer, nullable=False, unique=True)
+    telegram_id = sa.Column(sa.BigInteger, nullable=False, unique=True)
     fullname = sa.Column(sa.String)
     username = sa.Column(sa.String)
